+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 16.11.2015
 ==========
 
+* Add method `ULogin::getUserAttributes($token)`.
+* Add action for response from ulogin.
 * Deprecated: renamed `Widget::$lang` to `Widget::$language`.
 * Deprecated: renamed `Widget::$redirect_uri` to `Widget::$redirectUri`.
 * Deprecated: renamed `Widget::$execute_custom_init` to `Widget::$executeCustomInit`.